:green_heart: break up storage api tests to individual files
